# example-base-05/md-markdown-viewer-client/src/types.ts
export interface MarkdownFileMeta {
  name: string;
  slug: string;
  relativePath: string;
  index: number;
  total: number;
}

export interface MarkdownFile extends MarkdownFileMeta {
  content: string;
}

export interface FilesResponse {
  total: number;
  files: MarkdownFileMeta[];
}