# example-base-05/md-markdown-viewer/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Markdown Explorer API",
    "description": "Simple API for inspecting Markdown discovery status.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3030",
      "description": "Local dev server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Render markdown viewer UI",
        "description": "Renders the EJS user interface. Use the `slug` query parameter to select a specific Markdown file.",
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": false,
            "description": "Slug of the markdown file to highlight.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "description": "When true, forces a rescan of markdown folders before rendering.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML content"
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health status",
        "responses": {
          "200": {
            "description": "Health payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    },
                    "totalFiles": {
                      "type": "integer",
                      "example": 42
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/files": {
      "get": {
        "summary": "List markdown files",
        "description": "Returns metadata for every markdown file discovered on disk.",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "description": "When true, forces a rescan of markdown folders before responding.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of markdown files",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total": {
                      "type": "integer",
                      "example": 2
                    },
                    "files": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MarkdownFile"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{slug}": {
      "get": {
        "summary": "Get markdown file by slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Markdown file with content",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/MarkdownFile"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "content": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "File not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MarkdownFile": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "intro"
          },
          "slug": {
            "type": "string",
            "example": "sample-markdown-intro-1a2b3c"
          },
          "relativePath": {
            "type": "string",
            "example": "intro.md"
          },
          "index": {
            "type": "integer",
            "example": 1
          },
          "total": {
            "type": "integer",
            "example": 5
          }
        }
      }
    }
  }
}